+++ CHANGELOG.md
@@ -34,6 +34,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#3965](https://github.com/open-telemetry/opentelemetry-python/pull/3965))
 - Validate links at span creation
   ([#3991](https://github.com/open-telemetry/opentelemetry-python/pull/3991))
+- Fix to add attributes field in get_meter and instrumentationScope
+  ([#3695](https://github.com/open-telemetry/opentelemetry-python/issues/3695))
 
 ## Version 1.25.0/0.46b0 (2024-05-30)
 
@@ -1601,6 +1603,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#3778](https://github.com/open-telemetry/opentelemetry-python/pull/3778))
 - Fix license field in pyproject.toml files
   ([#3803](https://github.com/open-telemetry/opentelemetry-python/pull/3803))
-- Fix to add attributes field in get_meter and instrumentationScope
-  ([#3695](https://github.com/open-telemetry/opentelemetry-python/issues/3695))